Add timer cancel feature to test
